pyproject.toml
[project]
name = "nlp-project"
version = "0.1.0"
description = ""
authors = [
    {name = "avolxn",email = "vlad.eldashov2303@mail.ru"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "langchain (>=0.3.25,<0.4.0)",
    "langchain-openai (>=0.3.16,<0.4.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-huggingface (>=0.2.0,<0.3.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pypdf (>=5.5.0,<6.0.0)",
    "docx2txt (>=0.9,<0.10)",
    "unstructured (>=0.17.2,<0.18.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
